moved files to other catalogs, renamed method, changed blender_python_commands.py to enum
