+++ .gitignore
@@ -292,3 +292,4 @@ Lib/
 figures/
 Scripts/
 ../data/
+blender3d_intergration/object_trajectory_calculation/calculated_trajectories
